Resolution,LongFlag,Barnola
Decreased resolution plus smoothing and array extension

+++ Python/COD_read.py
@@ -14,8 +14,10 @@ import seaborn as sns
 #plt.rc('text', usetex=True)
 #plt.rc('font', family='serif')
 sns.set()
+from scipy.signal import savgol_filter
+from scipy.ndimage.filters import uniform_filter1d
 class CoD_plotter():
-    
+
     def __init__(self,filepath=None):
 
         self.filepath = filepath
@@ -43,16 +45,18 @@ class CoD_plotter():
             idx = int(np.where(self.z[i, 1:] == self.close_off_depth[i])[0])
             self.d15N_cod[i] = self.d15N[i,idx]
             self.d40Ar_cod[i] = self.d40Ar[i,idx]
-
-
-
+        
+        self.close_off_depth = savgol_filter(self.close_off_depth, window_length=51,polyorder=3)
+        self.d15N_cod = savgol_filter(self.d15N_cod, window_length=51, polyorder=3)
         #self.d15N_codm = np.mean(self.d15N_cod.reshape(-1,3), axis=1)
         #self.d40Ar_codm = np.mean(self.dAr40_cod.reshape(-1,3), axis=1)
         self.model_time[0] = 0    
         f.close()
-        
-        #plt.plot()
+       
         return
+    
+    
+    
     def plotting(self):
         rows,cols = 2,3
         self.fig,self.ax = plt.subplots(rows,cols,figsize=(15, 15), tight_layout=True)
@@ -65,8 +69,8 @@ class CoD_plotter():
         self.ax[1,0].set_xlim(self.model_time[0],self.model_time[-1])
         self.ax[1,2].set_xlim(self.model_time[0],self.model_time[-1])
         
-        Times = np.array([2000,4000,5000,6000,8000,9900,10000-1])
-        Times = [500,1000,2000,3000,4000,4700,5000-1]
+        Times = [500,2000,4000,6000,8000,9900,9990-1]
+        #Times = [500,1000,2000,3000,4000,4700,5000-1]
         #Times = [250,500,1000,1500,2000,2350]
         cmap_interval = np.linspace(0,1,len(Times))
         time_labels = ['$t_0$', '$t_1$', '$t_2$', '$t_3$', '$t_4$', '$t_5$', '$t_6$','$t_f$']
@@ -74,7 +78,7 @@ class CoD_plotter():
         
         self.ax02 = self.ax[0,0].twinx()
         self.ax[0,0].set_ylabel("Temperature Forcing [K]",color=cmap(cmap_interval[3]),fontsize=labelFont)
-        self.ax02.set_ylabel("Accumulation Forcing [$\mathrm{my}^{-1}$ ice eq.]",color=cmap(cmap_interval[-1]),fontsize=labelFont)
+        self.ax02.set_ylabel("Accumulation Forcing [$\mathrm{my}^{-1}$ ice eq.]",color=cmap(cmap_interval[1]),fontsize=labelFont)
         self.ax[0,0].set_xlabel("Model-time [yr]",fontsize=labelFont)
         #self.ax[0,0].vlines(Times, 0,1 , transform=self.ax[0,0].get_xaxis_transform(), colors='b',linestyles='dashed',alpha=0.5)
 
@@ -109,9 +113,10 @@ class CoD_plotter():
         self.ax[1,1].grid(linestyle=':', color='gray', lw='0.3')
         self.ax[1,2].grid(linestyle=':', color='gray', lw='0.3')
         
-        
+        #print(Times)
+        #print(self.d15N.shape)
         self.ax[0,0].plot(self.climate[:,0],self.climate[:,2],color=cmap(cmap_interval[3]))
-        self.ax02.plot(self.climate[:,0],self.climate[:,1],color=cmap(cmap_interval[-1]))
+        self.ax02.plot(self.climate[:,0],self.climate[:,1],color=cmap(cmap_interval[1]))
         for i in range(len(Times)):
             self.ax[0,1].plot(self.rho[Times[i]][1:], self.z[Times[i]][1:],color=cmap(cmap_interval[i]),linestyle='--',linewidth=0.7,label=time_labels[i])
             self.ax[0,2].plot(self.d15N[Times[i]][1:], self.z[Times[i]][1:],color=cmap(cmap_interval[i]),linestyle='-',linewidth=0.7,label=time_labels[i])
@@ -139,10 +144,12 @@ class CoD_plotter():
         
         
         #print(self.model_time[1],self.model_time.min())
-'''                     
-Test = CoD_plotter(filepath='CFM/CFM_main/CFMoutput/Temp_linear')
-Test.plotting()
-'''
+                    
+
+
+
+
+
 
 folder = './CFM/CFM_main/CFMinput'
 
@@ -150,22 +157,15 @@ Folder = [name for name in os.listdir(folder) if os.path.isdir(os.path.join(fold
 rfolder = 'CFM/CFM_main/CFMoutput/'
 
 for i in range(len(Folder)):
-    print(i)
-    Current_plot = CoD_plotter(filepath=rfolder+Folder[i])
-    Current_plot.plotting()
-    plt.savefig('CoD_plots/'+ str(Folder[i])+'.png',dpi=300)
-
+    print(Folder[i])
+    if os.path.exists(rfolder + Folder[i]+"/CFMresults.hdf5"):
+        Current_plot = CoD_plotter(filepath=rfolder+Folder[i])
+        Current_plot.plotting()
+        plt.savefig('CoD_plots/'+ str(Folder[i])+'.png',dpi=300)
+        plt.close('all')
+    else:
+        print("Results file does not exist")
 plt.close('all')
-
-
-'''
-for i in range(len(Folder)):
-    print(i)
-    Current_plot = CoD_plotter(filepath=FlipFolder+Folder[i])
-    Current_plot.plotting()
-    plt.savefig('CoD_plots/Flip'+ str(Folder[i])+'.png',dpi=300)
-
-'''
